string-bytes: Fctor functions that convert string to byte representations out of the common module.
These functions are very much self contained. Having them mixed in with
the common modules makes porting the tools over to the new options /
context stuff.
