+++ Makefile.am
@@ -48,8 +48,9 @@ COMMON_SRC = \
     src/context-util.c \
     src/files.c \
     src/options.c \
-    src/tpm2-header.c \
-    src/log.c
+    src/log.c \
+    src/string-bytes.c \
+    src/tpm2-header.c
 
 # The tools within this conditional are hard coded to the socket TCTI.
 # As they're migrated to support dynamic TCTI configuration they should be
